moved api calling functions to their own folder for easier testing
