+++ cmd/local/apicollect/testdata/conf/dremio.conf
@@ -1,0 +1,35 @@
+#
+# Copyright (C) 2017-2019 Dremio Corporation
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#     http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+#
+
+paths: {
+  # the local path for dremio to store data.
+  local: ${DREMIO_HOME}"/data"
+
+  # the distributed path Dremio data including job results, downloads, uploads, etc
+  #dist: "pdfs://"${paths.local}"/pdfs"
+}
+
+services: {
+  coordinator.enabled: true,
+  coordinator.master.enabled: true,
+  executor.enabled: true,
+  flight.use_session_service: true
+}
+
+
+debug: {
+    addDefaultUser: true
+}
