+++ cmd/local/apicollect/validate.go
@@ -1,0 +1,30 @@
+//	Copyright 2023 Dremio Corporation
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//	http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+// apicollect provides all the methods that collect via the API, this is a substantial part of the activities of DDC so it gets it's own package
+package apicollect
+
+import (
+	"github.com/dremio/dremio-diagnostic-collector/cmd/local/conf"
+	"github.com/dremio/dremio-diagnostic-collector/cmd/local/restclient"
+	"github.com/dremio/dremio-diagnostic-collector/cmd/simplelog"
+)
+
+func ValidateAPICredentials(c *conf.CollectConf) error {
+	simplelog.Info("Validating REST API user credentials...")
+	url := c.DremioEndpoint() + "/apiv2/login"
+	headers := map[string]string{"Content-Type": "application/json"}
+	_, err := restclient.APIRequest(url, c.DremioPATToken(), "GET", headers)
+	return err
+}
